CyberOdyssey-2024-Finals/FlagSideQuest/challenge/requirements.txt
discord.py
PyJWT[crypto]
cryptography
pycryptodome